refactor: finish the C23 migration and fix the bugs it uncovered
The extension was half through the move from C++ to C23, and the two halves
disagreed. Strings crossed the C boundary as a bare `char *`, so a value that
carried a NUL byte parsed as its prefix: `new Cassandra\Inet("127.0.0.1\0x")`
built `127.0.0.1`, and the same held for every UUID and every numeric string.
The `-1` that `Time`, `Date` and `Timeuuid` used to mean "no argument" was also
a value a caller could pass, so `new Cassandra\Time(-1)` returned the current
time of day instead of a rejection. All 27 property handlers replaced the
object property table on every call, and because the engine keys an iterator
position on the table address, a `foreach` whose body read the same object
never reached the end.

Strings are `zend_string` throughout now, the driver calls take the
length-aware `_n` form, and every message that carries a string from PHP is
bounded by its real length. The constructors read the argument count instead of
a sentinel, the numeric parser requires a digit after the sign, and a property
handler empties the table it already owns. `Type\UserType` compares its
keyspace and name, so `ks.x` and `ks.y` are no longer the same type. The build
enables a much wider warning set and `src/` compiles clean under it on GCC and
on Clang, with sanitizer, GCC analyzer and clang-tidy runs clean.

+++ include/php_scylladb_cache_key.h
@@ -58,7 +58,7 @@ static zend_always_inline zend_ulong php_scylladb_cache_key_mix_long(zend_ulong 
 }
 
 [[gnu::pure]]
-static zend_always_inline zend_ulong php_scylladb_cache_key_mix_int(zend_ulong h, int v) {
+static zend_always_inline zend_ulong php_scylladb_cache_key_mix_int(zend_ulong h, int64_t v) {
     return php_scylladb_cache_key_mix_bytes(h, &v, sizeof(v));
 }
